+++ app/src/components/pages/searchPage/searchPage.tsx
@@ -9,13 +9,15 @@ import {
   TagSet,
   Link,
   Icon,
+  Menu,
 } from "@nypl/design-system-react-components";
 import React, { useRef } from "react";
-import { CARDS_PER_PAGE } from "@/src/config/constants";
+import { CARDS_PER_PAGE, SEARCH_SORT_LABELS } from "@/src/config/constants";
 import { displayResults, totalNumPages } from "@/src/utils/utils";
 import Filters from "../../search/filters/filters";
 import { useSearchContext } from "@/src/context/SearchProvider";
 import { usePathname, useRouter } from "next/navigation";
+import SearchCardsGrid from "../../grids/searchCardsGrid";
 
 const SearchPage = ({ data }) => {
   const { searchManager } = useSearchContext();
@@ -66,7 +68,6 @@ const SearchPage = ({ data }) => {
           paddingRight: { base: "m", xl: "s" },
         }}
       >
-        <HorizontalRule />
         <Flex alignContent="center" gap="xs">
           <Text size="subtitle2" sx={{ margin: 0, fontWeight: 400 }}>
             Filters applied:
@@ -83,18 +84,42 @@ const SearchPage = ({ data }) => {
           />
         </Flex>
         <HorizontalRule />
-        <Heading
-          ref={headingRef}
-          tabIndex={-1}
-          marginTop="xl"
-          size="heading5"
-        >{`Displaying ${displayResults(
-          data.numResults,
-          CARDS_PER_PAGE,
-          searchManager.page
-        )}
-                    results`}</Heading>
-
+        <Flex
+          justifyContent="space-between"
+          sx={{
+            marginTop: { base: "m", md: "xl" },
+            marginBottom: { base: "m", md: "s" },
+            flexDir: { base: "column", md: "row" },
+            gap: "m",
+            alignItems: { base: "flex-start", md: "center" },
+          }}
+        >
+          <Heading
+            ref={headingRef}
+            size="heading5"
+            tabIndex={-1}
+            margin="0"
+          >{`Displaying ${displayResults(
+            data.numResults,
+            CARDS_PER_PAGE,
+            searchManager.page
+          )}
+                                        results`}</Heading>
+          <Menu
+            showLabel
+            selectedItem={"relevance"}
+            labelText={`Sort by: ${SEARCH_SORT_LABELS["relevance"]}`}
+            listItemsData={Object.entries(SEARCH_SORT_LABELS).map(
+              ([id, label]) => ({
+                id,
+                label,
+                onClick: () => {},
+                type: "action",
+              })
+            )}
+          />
+        </Flex>
+        <SearchCardsGrid keywords={data.keyword} results={data.results} />
         <Flex marginTop="xxl" marginBottom="xxl" alignContent="center">
           <Link
             minWidth="100px"